+++ Sources/RichTextKit/Extensions/NSImage+CGImage.swift
@@ -1,0 +1,18 @@
+//
+//  NSImage+JpegData.swift
+//  RichTextKit
+//
+//  Created by Daniel Saidi on 2022-06-05.
+//  Copyright © 2022 Daniel Saidi. All rights reserved.
+//
+
+#if canImport(AppKit)
+import AppKit
+
+public extension NSImage {
+
+    var cgImage: CGImage? {
+        cgImage(forProposedRect: nil, context: nil, hints: nil)
+    }
+}
+#endif